+++ README.md
@@ -1,5 +1,22 @@
 # Okta .NET SDK
 
-This branch contains **work-in-progress** for version 1.0 of the Okta .NET SDK. This version is a major breaking change from all previous versions.
+This repository contains the Okta Management SDK for .NET (C# and Visual Basic). This SDK can be used in your server code to create and update users, manage groups and roles, and more.
 
-For the currently-released stable version, see the [package on NuGet](https://www.nuget.org/packages/Okta.Core.Client).
+The SDK is compatible with:
+* [.NET Standard](https://docs.microsoft.com/en-us/dotnet/standard/library) 1.4
+* .NET Framework 4.6.1
+* Mono
+
+## Development version (1.x)
+
+The [master](https://github.com/okta/oktasdk-csharp/tree/master) branch contains the 1.x version of the SDK, which is currently under active development. This version is a major update, and is a breaking change from all previously-released packages. It has not yet been published to NuGet.
+
+If you want to be notified when updates are released, use the Watch button above to watch this repository!
+
+## Previous version (0.x)
+
+The [`0.x` branch](https://github.com/okta/oktasdk-csharp/tree/0.x) contains the previous version of the SDK. This is published on NuGet as [Okta.Core.Client 0.3.3](https://www.nuget.org/packages/Okta.Core.Client/0.3.3) and lower.
+
+## Getting help
+
+This library is maintained and supported by Okta. If you run into trouble using the SDK, post an [issue](https://github.com/okta/oktasdk-csharp/issues) or send us an email at [developers@okta.com](mailto:developers@okta.com).
