--- THIRD-PARTY-NOTICES
@@ -1,41 +1,0 @@
-THIRD PARTY NOTICES
-
-This document contains third party open source licenses and notices for
-the Okta oktasdk-csharp product. Certain licenses and notices may appear
-in other parts of the product in accordance with the applicable license
-requirements.
-
-The Okta product that this document references does not necessarily use
-all the open source software packages referred to below and may also
-only use portions of a given package.
-
-________________________
-
-Json.NET 6.0.6
-http://www.nuget.org/packages/Newtonsoft.Json/
-https://raw.githubusercontent.com/JamesNK/Newtonsoft.Json/master/LICENSE.md
-
-Microsoft HTTP Client Libraries 2.2.28
-http://www.nuget.org/packages/Microsoft.Net.Http/
-http://www.microsoft.com/net/dotnet_library_license.htm
-
-________________________
-
-Okta makes the following disclaimers regarding the Third Party Components on
-behalf of itself, the copyright holders, contributors, and licensors of such
-Third Party Components:
-
-TO THE FULLEST EXTENT PERMITTED UNDER APPLICABLE LAW, THE THIRD PARTY
-COMPONENTS ARE PROVIDED BY THE COPYRIGHT HOLDERS, CONTRIBUTORS, LICENSORS, AND
-OKTA "AS IS" AND ANY REPRESENTATIONS OR WARRANTIES OF ANY KIND, WHETHER ORAL
-OR WRITTEN, WHETHER EXPRESS, IMPLIED, OR ARISING BY STATUTE, CUSTOM, COURSE OF
-DEALING, OR TRADE USAGE, INCLUDING WITHOUT LIMITATION THE IMPLIED WARRANTIES
-OF TITLE, MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, AND
-NON-INFRINGEMENT, ARE DISCLAIMED. IN NO EVENT WILL THE COPYRIGHT OWNER,
-CONTRIBUTORS, LICENSORS, OR OKTA BE LIABLE FOR ANY DIRECT, INDIRECT,
-INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA,
-OR PROFITS; OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY OF
-LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE
-OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THE THIRD PARTY COMPONENTS,
-EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
