--- src/Okta.Sdk.Abstractions/ISaveable.cs
@@ -1,14 +1,0 @@
-﻿// <copyright file="ISaveable.cs" company="Okta, Inc">
-// Copyright (c) 2014-2017 Okta, Inc. All rights reserved.
-// Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
-// </copyright>
-
-using System.Collections.Generic;
-
-namespace Okta.Sdk.Abstractions
-{
-    public interface ISaveable
-    {
-        IReadOnlyDictionary<string, object> GetModifiedProperties();
-    }
-}
