+++ src/Okta.Sdk.UnitTests/DefaultSerializerShould.cs
@@ -6,6 +6,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using FluentAssertions;
+using Okta.Sdk.Internal;
 using Xunit;
 
 namespace Okta.Sdk.UnitTests
