+++ src/Okta.Sdk.UnitTests/MockedCollectionRequestExecutor{T}.cs
@@ -9,7 +9,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
-using Okta.Sdk.Abstractions;
+using Okta.Sdk.Internal;
 
 namespace Okta.Sdk.UnitTests
 {
