+++ src/Okta.Sdk.UnitTests/ResourceShould.cs
@@ -6,6 +6,7 @@
 using System;
 using System.Collections.Generic;
 using FluentAssertions;
+using Okta.Sdk.Internal;
 using Xunit;
 
 namespace Okta.Sdk.UnitTests
