+++ src/Okta.Sdk.UnitTests/TestResource.cs
@@ -3,6 +3,8 @@
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 // </copyright>
 
+using Okta.Sdk.Internal;
+
 namespace Okta.Sdk.UnitTests
 {
     public class TestResource : Resource
