+++ src/Okta.Sdk/CollectionClient{T}.cs
@@ -5,7 +5,7 @@
 
 using System;
 using System.Collections.Generic;
-using Okta.Sdk.Abstractions;
+using Okta.Sdk.Internal;
 
 namespace Okta.Sdk
 {
