+++ src/Okta.Sdk/Configuration/CacheManagerConfiguration.cs
@@ -3,7 +3,7 @@
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 // </copyright>
 
-namespace Okta.Sdk.Abstractions
+namespace Okta.Sdk.Configuration
 {
     public sealed class CacheManagerConfiguration
     {
