+++ src/Okta.Sdk/IOktaClient.cs
@@ -5,7 +5,7 @@
 
 using System.Threading;
 using System.Threading.Tasks;
-using Okta.Sdk.Abstractions;
+using Okta.Sdk.Internal;
 
 namespace Okta.Sdk
 {
