+++ src/Okta.Sdk/Internal/DefaultChangeTrackingDictionary.cs
@@ -7,9 +7,8 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using Okta.Sdk.Abstractions;
 
-namespace Okta.Sdk
+namespace Okta.Sdk.Internal
 {
     public sealed class DefaultChangeTrackingDictionary : IChangeTrackingDictionary<string, object>
     {
