+++ src/Okta.Sdk/Internal/DefaultRequestExecutor.cs
@@ -10,9 +10,8 @@ using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
-using Okta.Sdk.Abstractions;
 
-namespace Okta.Sdk
+namespace Okta.Sdk.Internal
 {
     public sealed class DefaultRequestExecutor : IRequestExecutor
     {
