+++ src/Okta.Sdk/Internal/DefaultSerializer.cs
@@ -6,9 +6,8 @@
 using System.Collections.Generic;
 using System.IO;
 using Newtonsoft.Json;
-using Okta.Sdk.Abstractions;
 
-namespace Okta.Sdk
+namespace Okta.Sdk.Internal
 {
     public sealed class DefaultSerializer : ISerializer
     {
