+++ src/Okta.Sdk/Internal/IDataStore.cs
@@ -6,9 +6,8 @@
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
-using Okta.Sdk.Abstractions;
 
-namespace Okta.Sdk
+namespace Okta.Sdk.Internal
 {
     public interface IDataStore
     {
