+++ src/Okta.Sdk/Internal/ISerializer.cs
@@ -5,7 +5,7 @@
 
 using System.Collections.Generic;
 
-namespace Okta.Sdk.Abstractions
+namespace Okta.Sdk.Internal
 {
     public interface ISerializer
     {
