+++ src/Okta.Sdk/Internal/UrlFormatter.cs
@@ -3,13 +3,11 @@
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 // </copyright>
 
-using System;
 using System.Linq;
 using System.Net;
 using System.Text;
-using Okta.Sdk.Abstractions;
 
-namespace Okta.Sdk
+namespace Okta.Sdk.Internal
 {
     public static class UrlFormatter
     {
