+++ src/Okta.Sdk/OktaApiException.cs
@@ -3,7 +3,6 @@
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 // </copyright>
 
-using Okta.Sdk.Abstractions;
 
 namespace Okta.Sdk
 {
