+++ src/Okta.Sdk/OktaClient.cs
@@ -3,12 +3,11 @@
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 // </copyright>
 
-using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Logging.Abstractions;
-using Okta.Sdk.Abstractions;
+using Okta.Sdk.Configuration;
+using Okta.Sdk.Internal;
 
 namespace Okta.Sdk
 {
