+++ src/Okta.Sdk/Resource.cs
@@ -8,6 +8,7 @@ using System.Collections.Generic;
 using System.Reflection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
+using Okta.Sdk.Internal;
 
 namespace Okta.Sdk
 {
