+++ todo.md
@@ -1,9 +1,7 @@
 # v1.0 todo
 
-* Generate model interfaces
 * Handle crud methods?
 * Consider moving internal stuff to a new namespace
 * Update generator to public version
-* Add string enumerations
 * Remove `Generated/` from `.gitignore`
 * Fix extra whitespace in generated code (low prio)
